pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "fridgeos"
version = "0.1.2"
description = "Simple, easy-to-use control software for cryostats"
authors = [
    { name = "Adam McCaughan", email = "adam.mccaughan@nist.gov"},
    { name = "Ryan Morgenstern", email = "ryan.morgenstern@nist.gov"},
]
license = { text = "CC BY-SA" }
dependencies = [
"psycopg2-binary",
"pandas",
"requests",
"pyserial",
"pyzmq",
]

#[tool.setuptools]
#packages = ["hal", "monitor", "scraper", "statemachine", "zmqhelper"]

#[tool.setuptools.packages.find]
#include = ["hal", "monitor", "scraper", "statemachine", "zmqhelpers"]
#namespaces = false
